backend/anfu_backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
cryptography==46.0.2
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
geopandas==1.1.1
gunicorn==23.0.0
idna==3.10
numpy==1.26.4
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
pyogrio==0.11.1
pyproj==3.6.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
shapely==2.1.2
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
psycopg2-binary
scikit-learn>=1.3.0
pandas>=2.1.0


